refactor: 메소드의 클래스 위치 변경 #129
