+++ server/src/main/java/com/hobbyt/domain/member/service/MemberService.java
@@ -4,28 +4,21 @@ import static com.hobbyt.domain.member.entity.MemberStatus.*;
 import static com.hobbyt.global.error.exception.ExceptionCode.*;
 import static com.hobbyt.global.security.constants.AuthConstants.*;
 
-import java.util.List;
-
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.hobbyt.domain.follow.repository.FollowRepository;
-import com.hobbyt.domain.member.dto.request.ProfileRequest;
 import com.hobbyt.domain.member.dto.request.SignupRequest;
 import com.hobbyt.domain.member.dto.request.UpdateMyInfoRequest;
 import com.hobbyt.domain.member.dto.request.UpdatePassword;
 import com.hobbyt.domain.member.dto.response.MyInfoResponse;
-import com.hobbyt.domain.member.dto.response.ProfileResponse;
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.domain.member.entity.Recipient;
 import com.hobbyt.domain.member.repository.MemberRepository;
-import com.hobbyt.domain.privatehome.service.PrivateHomeService;
 import com.hobbyt.global.entity.Account;
 import com.hobbyt.global.error.exception.BusinessLogicException;
 import com.hobbyt.global.redis.RedisService;
 import com.hobbyt.global.security.jwt.JwtTokenProvider;
-import com.hobbyt.global.security.member.MemberDetails;
 
 import lombok.RequiredArgsConstructor;
 
@@ -37,8 +30,6 @@ public class MemberService {
 	private final PasswordEncoder passwordEncoder;
 	private final JwtTokenProvider jwtTokenProvider;
 	private final RedisService redisService;
-	private final FollowRepository followRepository;
-	private final PrivateHomeService privateHomeService;
 	private final String path = "/api/images/";
 	private final String defaultProfileImage = "a30a68de-0bab-45c0-93ec-1802de8c62ed.jpg";
 	private final String defaultHeaderImage = "e048f178-9a96-4f59-a6e9-8991abb700d7.jpg";
@@ -148,47 +139,8 @@ public class MemberService {
 		return MyInfoResponse.of(member);
 	}
 
-	@Transactional
-	public ProfileResponse getProfile(final Long targetMemberId, MemberDetails loginMember) {
-		Member targetMember = findMemberById(targetMemberId);
-		ProfileResponse profileResponse = null;
-
-		if (loginMember == null) {
-			profileResponse = ProfileResponse.of(targetMember);
-			return profileResponse;
-		}
-
-		String email = loginMember.getEmail();
-		privateHomeService.countVisitor(targetMemberId, email);
-		profileResponse = ProfileResponse.of(targetMember);
-		profileResponse.setIsFollowing(isFollowing(targetMemberId, email));
-
-		return profileResponse;
-	}
-
-	private Boolean isFollowing(Long targetMemberId, String email) {
-		Member myInfo = findMemberByEmail(email);
-
-		if (targetMemberId != myInfo.getId()) {
-			List<Long> myFollowingId = followRepository.findFollowingIdByMember(myInfo);
-			return myFollowingId.contains(targetMemberId);
-		}
-
-		return null;
-	}
-
 	public Member findMemberById(Long id) {
 		return memberRepository.findById(id)
 			.orElseThrow(() -> new BusinessLogicException(MEMBER_NOT_FOUND));
-	}
-
-	@Transactional
-	public void updateProfile(final String email, final ProfileRequest profileRequest) {
-
-		Member member = findMemberByEmail(email);
-
-		Member updateProfile = profileRequest.toEntity();
-
-		member.updateProfile(updateProfile);
 	}
 }
