+++ server/src/main/java/com/hobbyt/domain/privatehome/service/PrivateHomeService.java
@@ -1,13 +1,17 @@
 package com.hobbyt.domain.privatehome.service;
 
-import static com.hobbyt.global.error.exception.ExceptionCode.*;
+import java.util.List;
 
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.hobbyt.domain.follow.repository.FollowRepository;
+import com.hobbyt.domain.member.dto.request.ProfileRequest;
+import com.hobbyt.domain.member.dto.response.ProfileResponse;
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.domain.member.repository.MemberRepository;
+import com.hobbyt.domain.member.service.MemberService;
 import com.hobbyt.domain.privatehome.dto.PrivateHomeCommentResponse;
 import com.hobbyt.domain.privatehome.dto.PrivateHomePostLikeResponse;
 import com.hobbyt.domain.privatehome.dto.PrivateHomePostResponse;
@@ -16,7 +20,6 @@ import com.hobbyt.domain.privatehome.dto.PrivateHomeSaleLikeResponse;
 import com.hobbyt.domain.privatehome.dto.PrivateHomeSaleResponse;
 import com.hobbyt.domain.privatehome.entity.Visit;
 import com.hobbyt.domain.privatehome.repository.VisitRepository;
-import com.hobbyt.global.error.exception.BusinessLogicException;
 import com.hobbyt.global.security.member.MemberDetails;
 
 import lombok.RequiredArgsConstructor;
@@ -26,8 +29,9 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class PrivateHomeService {
 	private final MemberRepository memberRepository;
-	// private final MemberService memberService;
+	private final MemberService memberService;
 	private final VisitRepository visitRepository;
+	private final FollowRepository followRepository;
 
 	@Transactional
 	@Scheduled(cron = "0 0 0 * * *")
@@ -38,18 +42,53 @@ public class PrivateHomeService {
 
 	@Transactional
 	public void countVisitor(Long targetId, String visitorEmail) {
-		// Member visitor = memberService.findMemberByEmail(visitorEmail);
-		// Member target = memberService.findMemberById(targetId);
-		Member visitor = memberRepository.findByEmail(visitorEmail)
-			.orElseThrow(() -> new BusinessLogicException((MEMBER_NOT_FOUND)));
-		Member target = memberRepository.findById(targetId)
-			.orElseThrow(() -> new BusinessLogicException(MEMBER_NOT_FOUND));
+		Member visitor = memberService.findMemberByEmail(visitorEmail);
+		Member target = memberService.findMemberById(targetId);
 
 		if (canIncreaseVisitingNumber(visitor, target)) {
 			target.increaseVisitors();
 			Visit visit = new Visit(visitor, target);
 			visitRepository.save(visit);
 		}
+	}
+
+	@Transactional
+	public ProfileResponse getProfile(final Long targetMemberId, MemberDetails loginMember) {
+		Member targetMember = memberService.findMemberById(targetMemberId);
+		ProfileResponse profileResponse = null;
+
+		if (loginMember == null) {
+			profileResponse = ProfileResponse.of(targetMember);
+			return profileResponse;
+		}
+
+		String email = loginMember.getEmail();
+		countVisitor(targetMemberId, email);
+		profileResponse = ProfileResponse.of(targetMember);
+		profileResponse.setIsFollowing(isFollowing(targetMemberId, email));
+
+		return profileResponse;
+	}
+
+	@Transactional
+	public void updateProfile(final String email, final ProfileRequest profileRequest) {
+
+		Member member = memberService.findMemberByEmail(email);
+
+		Member updateProfile = profileRequest.toEntity();
+
+		member.updateProfile(updateProfile);
+	}
+
+	private Boolean isFollowing(Long targetMemberId, String email) {
+		Member myInfo = memberService.findMemberByEmail(email);
+
+		if (targetMemberId != myInfo.getId()) {
+			List<Long> myFollowingId = followRepository.findFollowingIdByMember(myInfo);
+			return myFollowingId.contains(targetMemberId);
+		}
+
+		return null;
 	}
 
 	private boolean canIncreaseVisitingNumber(Member visitor, Member target) {
@@ -65,13 +104,7 @@ public class PrivateHomeService {
 	}
 
 	@Transactional
-	public PrivateHomePostResponse getBlogListByMemberId(Long memberId, PrivateHomeRequest params,
-		MemberDetails loginMember) {
-
-		/*if (loginMember != null) {
-			countVisitor(memberId, loginMember.getEmail());
-		}*/
-
+	public PrivateHomePostResponse getBlogListByMemberId(Long memberId, PrivateHomeRequest params) {
 		return memberRepository.getBlogListByWriterId(memberId, params);
 	}
